clients/client-py/pyproject.toml
[tool.poetry]
name = "asst-client-py"
version = "0.1.0"
description = "Example of Python Client for the ASST project"
authors = ["nickosh <nikolay.shishov@gmail.com>"]
license = "MPL-2.0"
readme = "README.md"
packages = [{include = "asst_client_py"}]

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = {version = "5.7.1", extras = ["client"]}


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
